+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 严格遵循 [`Semantic Versioning 2.0.0`](http://semver.org/lang/zh-CN/) 语义化版本规范。
 
-## 5.2.3
+## 5.2.3(未发版)
 - @uiw-admin/basic-layouts
+ - 🌟 `basic-layouts`：增加`layouts` props,增加`useLayouts` hook
   - 🌟 `basic-layouts`：优化右上角menu关闭功能,通过useLayouts
+  - 🌟 `basic-layouts`：优化组件路径
 ## 5.2.1
 - @uiw-admin/basic-layouts
   - 🌟 `basic-layouts`：增加右上角菜单关闭功能